model_server/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
